Register runner with a temporary token (#5)
This version changes how a self-hosted runner is registered in GitHub.
A lambda function obtains a temporary registration token and puts it in
a secret. The instance has a permission to read the secret. The same
time, the instance doesn't need access to the permanent token. The
module revokes a read permission on the permanent token.

Co-authored-by: Oleksandr Kuzminskyi <aleksandr.kuzminsky@gmail.com>