Add autoscaling_scaleout_evaluation_period. Remove the GitHub registration token sooner
Before, the ASG waited a hardcoded one minute before scaling out. Add a
variable `autoscaling_scaleout_evaluation_period` with the default of 60
seconds to tune this. It's useful for sporadic short-lived jobs. We
don't want to launch an instance when it won't be needed anyway.

As soon as a runner is registered, the registration token is not needed.
The lambda functions will try to remove it.

* The registration lambda will remove the token when the runner is
  known. This can happen when the instance comes from hibernation.

* The deregistration lambda will delete the token if it's still around.